Add multiple interop views

+++ compose/mpp/demo-uikit/src/uikitMain/kotlin/androidx/compose/mpp/demo/SwiftHelper.kt
@@ -19,5 +19,5 @@ package androidx.compose.mpp.demo
 import platform.UIKit.UIViewController
 
 class SwiftHelper {
-    fun getViewController(makeHostingViewController: () -> UIViewController): UIViewController = getViewControllerWithCompose(makeHostingViewController)
+    fun getViewController(makeHostingViewController: (Int) -> UIViewController): UIViewController = getViewControllerWithCompose(makeHostingViewController)
 }
