+++ compose/mpp/demo-uikit/src/uikitMain/kotlin/androidx/compose/mpp/demo/getViewControllerWithCompose.kt
@@ -21,6 +21,6 @@ import platform.UIKit.UIViewController
 
 // TODO This module is just a proxy to run the demo from mpp:demo. Figure out how to get rid of it.
 //  If it is removed, there is no available configuration in IDE
-fun getViewControllerWithCompose(makeHostingViewController: () -> UIViewController) = ComposeUIViewController {
+fun getViewControllerWithCompose(makeHostingViewController: (Int) -> UIViewController) = ComposeUIViewController {
     IosDemo("", makeHostingViewController)
 }
